requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
python-slugify
Pillow

